pyproject.toml
[project]
name = "lebot"
version = "0.1.0"
description = "A bot to help with manging cloud infra."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts==0.2.1",
    "discord-py==2.4.0",
    "mistralai==1.5.0",
    "python-dotenv==1.0.1",
    "aiohttp==3.11.12",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "attrs==25.1.0",
    "black==25.1.0",
    "boto3==1.37.0",
    "botocore==1.37.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "eval-type-backport==0.2.2",
    "frozenlist==1.5.0",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "jmespath==1.0.1",
    "jsonpath-python==1.0.6",
    "multidict==6.1.0",
    "mypy-extensions==1.0.0",
    "nba-api==1.7.0",
    "numpy==1.26.4",
    "packaging==24.2",
    "pathspec==0.12.1",
    "platformdirs==4.3.6",
    "propcache==0.2.1",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "python-dateutil==2.9.0.post0",
    "requests==2.32.3",
    "s3transfer==0.11.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "typing-extensions==4.12.2",
    "typing-inspect==0.9.0",
    "urllib3==2.3.0",
    "yarl==1.18.3",
]
